Port to x86_64: fix mppc putlit, pcre, cstring/iconv/climits/ctime, 64-bit property & tasks, Rules.make portability, CI build->release
- cnet/io/mppc.h putlit C rewrite (setne %h0 not encodable in 64-bit)
- provide pcre shim + libpcre dummy and -lpcre/-lcrypto handling for runners
- cnet/mk and cskill/mk: -m64 auto, -fpermissive, -include cstring/iconv/climits/ctime for modern g++
- cgame/Rules.make portable (abspath, -m64 -fpermissive -lcrypto) and libgs paths
- cgame/gs/* LP64 guards (TaskInterface ActiveTaskEntry 64B, BitImage, protocol_imp, iconv, typename)
- fix memcmp/time/INT_MAX/world_manager incomplete type (property.h) and GDB:: qualification
- add .github/workflows/build.yml (workflow_dispatch, x86_64, apt libssl/pcre, build cnet->cskill->cgame, package + upload + Release)

+++ cnet/gdbclient/db_if.h
@@ -151,7 +151,7 @@ struct base_info
 struct ivec
 {
 	const void * data;
-	unsigned int size;
+	size_t size;
 };
 
 struct itemdata
@@ -166,7 +166,7 @@ struct itemdata
 	int proctype;
 	int expire_date;
 	const void * data;
-	unsigned int size;
+	size_t size;
 };
